move lookUpFunction method to HasGrcService service
