+++ workers/loc.api/has.grc.service/index.js
@@ -1,5 +1,6 @@
 'use strict'
 
+const { promisify } = require('util')
 const {
   decorate,
   injectable,
@@ -13,23 +14,35 @@ class HasGrcService {
     this.rService = rService
   }
 
+  async lookUpFunction (service) {
+    const link = this.rService.ctx.grc_bfx.link
+    const lookup = promisify(link.lookup).bind(link)
+
+    try {
+      const res = await lookup(service)
+
+      return Array.isArray(res)
+        ? res.length
+        : 0
+    } catch (err) {
+      return 0
+    }
+  }
+
   async hasS3AndSendgrid () {
-    const countS3Services = await this.rService.lookUpFunction(
-      null,
-      { params: { service: 'rest:ext:s3' } }
+    const countS3Services = await this.lookUpFunction(
+      'rest:ext:s3'
     )
-    const countSendgridServices = await this.rService.lookUpFunction(
-      null,
-      { params: { service: 'rest:ext:sendgrid' } }
+    const countSendgridServices = await this.lookUpFunction(
+      'rest:ext:sendgrid'
     )
 
     return !!(countS3Services && countSendgridServices)
   }
 
   async hasGPGService () {
-    const countPGPServices = await this.rService.lookUpFunction(
-      null,
-      { params: { service: 'rest:ext:gpg' } }
+    const countPGPServices = await this.lookUpFunction(
+      'rest:ext:gpg'
     )
 
     return !!countPGPServices
